perception: add test for blob/common/syncedmem

+++ modules/perception/base/syncedmem_test.cc
@@ -85,14 +85,14 @@ TEST_F(SyncedMemoryTest, TestCPUWrite) {
   void* cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::HEAD_AT_CPU);
   memset(cpu_data, 1, mem.size());
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<char*>(cpu_data))[i], 1);
   }
   // do another round
   cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::HEAD_AT_CPU);
   memset(cpu_data, 2, mem.size());
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<char*>(cpu_data))[i], 2);
   }
 }
@@ -109,21 +109,21 @@ TEST_F(SyncedMemoryTest, TestGPURead) {
   // check if values are the same
   char* recovered_value = new char[10];
   cudaMemcpy(recovered_value, gpu_data, 10, cudaMemcpyDefault);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<char*>(recovered_value))[i], 1);
   }
   // do another round
   cpu_data = mem.mutable_cpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::HEAD_AT_CPU);
   memset(cpu_data, 2, mem.size());
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<char*>(cpu_data))[i], 2);
   }
   gpu_data = mem.gpu_data();
   EXPECT_EQ(mem.head(), SyncedMemory::SYNCED);
   // check if values are the same
   cudaMemcpy(recovered_value, gpu_data, 10, cudaMemcpyDefault);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<char*>(recovered_value))[i], 2);
   }
   delete[] recovered_value;
@@ -136,7 +136,7 @@ TEST_F(SyncedMemoryTest, TestGPUWrite) {
   cudaMemset(gpu_data, 1, mem.size());
   const void* cpu_data = mem.cpu_data();
   EXPECT_NE(cpu_data, nullptr);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<const char*>(cpu_data))[i], 1);
   }
   EXPECT_EQ(mem.head(), SyncedMemory::SYNCED);
@@ -146,7 +146,7 @@ TEST_F(SyncedMemoryTest, TestGPUWrite) {
   cudaMemset(gpu_data, 2, mem.size());
   cpu_data = mem.cpu_data();
   EXPECT_NE(cpu_data, nullptr);
-  for (int i = 0; i < mem.size(); ++i) {
+  for (size_t i = 0; i < mem.size(); ++i) {
     EXPECT_EQ((static_cast<const char*>(cpu_data))[i], 2);
   }
   EXPECT_EQ(mem.head(), SyncedMemory::SYNCED);
